rename check-if-up
