+++ test/setup/TestEnv.ts
@@ -5,7 +5,7 @@ import { mkdirSync, rmSync } from 'fs';
 import * as readline from 'readline';
 import { SevereServiceError } from 'webdriverio';
 import TestSettings from '../types/TestSettings.js';
-import checkIfUp from './checkIfUp.js';
+import checkIfUp from './check-if-up.js';
 import { makeTestSettings } from './makeTestSettings.js';
 
 declare global {
